Episode 04: Moves api into 'custom' folder to prepare for react-query
