+++ package.json
@@ -28,6 +28,7 @@
     "lodash": "4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-query": "3.26.0",
     "react-scripts": "4.0.3",
     "styled-components": "5.3.1",
     "typescript": "4.4.3",
